Fix bug in :str parameter
Currently, the documentation says :str "can use special sequence [|] (e.g. octop[us|i]) where string to the left of the | will be used when the list contains just one item and the string to the right will be used otherwise," but it uses the left of the | when the list contains one OR TWO items.

This change corrects the behavior to conform with the documentation.